Refactor: Remove old scripts and add new utility modules
